requirements.txt
kafka-python
requests
pyyaml
